fixing branch_salary year and month as int4 adding salary period create table query adding salary period calculation query

+++ SQL/branch_salary.sql
@@ -6,8 +6,8 @@
 
 CREATE TABLE mekari.branch_salary (
 	branch_id int4 NULL,
-	"year" float8 NULL,
-	"month" float8 NULL,
+	"year" int4 NULL,
+	"month" int4 NULL,
 	total_work_hour float8 NULL,
 	total_salary int8 NULL,
 	salary_per_hour float8 NULL
